Implement day 9

+++ advent-of-code/9/kotlin/rubengees/src/main/kotlin/com/rubengees/day9/Utils.kt
@@ -4,11 +4,11 @@ package com.rubengees.day9
 
 import kotlin.math.pow
 
-fun Int.digits(): Sequence<Int> {
+fun Long.digits(): Sequence<Long> {
     var current = this
 
     return sequence {
-        while (current != 0) {
+        while (current != 0L) {
             yield(current % 10)
 
             current /= 10
@@ -16,10 +16,10 @@ fun Int.digits(): Sequence<Int> {
     }
 }
 
-fun Iterable<Int>.mergeDigits() = this
+fun Iterable<Long>.mergeDigits() = this
     .mapIndexed { index, i ->
         i * (10.0.pow((index).toDouble())).toInt()
     }
     .sum()
 
-fun Sequence<Int>.mergeDigits() = this.toList().mergeDigits()
+fun Sequence<Long>.mergeDigits() = this.toList().mergeDigits()
